// src/amap.d.ts
    export interface CustomLayerOption {
        opacity?: number;
        visible?: boolean;
        zIndex?: number;
        zooms?: [number, number];
        render?: () => void;
        alwaysRender?: boolean;
    }